FIX: Fix GLFW

+++ vispy/app/backends/_libglfw.py
@@ -519,9 +519,9 @@ def glfwGetWindowPos(window):
 
 
 def glfwGetCursorPos(window):
-    xpos, ypos = c_int(0), c_int(0)
+    xpos, ypos = c_double(0), c_double(0)
     _glfw.glfwGetCursorPos(window, byref(xpos), byref(ypos))
-    return xpos.value, ypos.value
+    return int(xpos.value), int(ypos.value)
 
 
 def glfwGetWindowSize(window):
